removing custom node tsx file, adding customizations in style node
